fix: rename test suite and lib module
